Split out ClassifierWrapper from the container Split the container into two components. Add some initial documentation, with an example.
