+++ packages/app-project/src/screens/ClassifyPage/components/ClassifierWrapper/ClassifierWrapperContainer.js
@@ -1,11 +1,8 @@
-import Classifier from '@zooniverse/classifier'
 import { MobXProviderContext, observer } from 'mobx-react'
-import auth from 'panoptes-client/lib/auth'
-import { func, shape } from 'prop-types'
-import React, { Component, useContext } from 'react'
-import asyncStates from '@zooniverse/async-states'
-import { logToSentry } from '@helpers/logger'
-import ErrorMessage from './components/ErrorMessage'
+import { func, string, shape } from 'prop-types'
+import React, { useContext } from 'react'
+
+import ClassifierWrapper from './classifierWrapper'
 
 export function storeMapper(store) {
   const {
@@ -29,124 +26,57 @@ export function storeMapper(store) {
   })
 }
 
-function withStore(Component) {
-  function DecoratedComponent(props) {
-    const { store } = useContext(MobXProviderContext)
-    const {
-      collections,
-      mode,
-      project,
-      recents,
-      user,
-      yourStats
-    } = storeMapper(store)
-    return (
-      <Component
-        collections={collections}
-        mode={mode}
-  // We use a POJO here, as the `project` resource is also stored in a
-  // `mobx-state-tree` store in the classifier and an MST node can't be in two
-  // stores at the same time.
-        project={project.toJSON()}
-        recents={recents}
-        user={user}
-        yourStats={yourStats}
-        {...props}
-      />
-    )
-  }
-  return observer(DecoratedComponent)
-}
+/**
+  A wrapper for the Classifier component. Connects the classifier to the project store and is responsible for handling:
+  - classifier errors.
+  - updates to project recents on classification complete.
+  - updates to stored favourites,when the classification subject is favourited.
+  - updates to stored collections, when the classification subject is added to a collection.
 
-export function ClassifierWrapperContainer({
-  onAddToCollection = () => true,
-  authClient = auth,
-  collections,
-  mode,
-  project,
-  recents,
-  subjectID,
-  subjectSetID,
-  user,
-  workflowID,
-  yourStats
-}) {
-
-  function onCompleteClassification(classification, subject) {
-    yourStats.increment()
-    recents.add({
-      favorite: subject.favorite,
-      subjectId: subject.id,
-      locations: subject.locations
-    })
-  }
-
-  function onError(error, errorInfo={}) {
-    logToSentry(error, errorInfo)
-    console.error('Classifier error', error)
-  }
-
-  function onToggleFavourite(subjectId, isFavourite) {
-    if (isFavourite) {
-      collections.addFavourites([subjectId])
-    } else {
-      collections.removeFavourites([subjectId])
-    }
-  }
-
-  const somethingWentWrong = project.loadingState === asyncStates.error
-
-  if (somethingWentWrong) {
-    const { error } = project
-    const errorToMessage = error || new Error('Something went wrong')
-    return (
-      <ErrorMessage error={errorToMessage} />
-    )
-  }
-
-  if (user.loadingState === asyncStates.loading) {
-    return (
-      <p>
-        Signing in…
-      </p>
-    )
-  }
-
-  try {
-    if (project.loadingState === asyncStates.success) {
-      const key = user.id || 'no-user'
-      return (
-        <Classifier
-          authClient={authClient}
-          key={key}
-          mode={mode}
-          onAddToCollection={onAddToCollection}
-          onCompleteClassification={onCompleteClassification}
-          onError={onError}
-          onToggleFavourite={onToggleFavourite}
-          project={project}
-          subjectID={subjectID}
-          subjectSetID={subjectSetID}
-          workflowID={workflowID}
-        />
-      )
-    }
-  } catch (error) {
-    onError(error)
-    return (
-      <ErrorMessage error={error} />
-    )
-  }
-
+  ```jsx
+  <ClassifierWrapper
+    onAddToCollection={addToCollection}
+    subjectID={subjectID}
+    subjectSetID={subjectSetID}
+    workflowID={workflowID}
+  />
+  ```
+*/
+function ClassifierWrapperContainer(props) {
+  const { store } = useContext(MobXProviderContext)
+  const {
+    collections,
+    mode,
+    project,
+    recents,
+    user,
+    yourStats
+  } = storeMapper(store)
   return (
-    <div>Loading…</div>
+    <ClassifierWrapper
+      collections={collections}
+      mode={mode}
+// We use a POJO here, as the `project` resource is also stored in a
+// `mobx-state-tree` store in the classifier and an MST node can't be in two
+// stores at the same time.
+      project={project.toJSON()}
+      recents={recents}
+      user={user}
+      yourStats={yourStats}
+      {...props}
+    />
   )
 }
 
 ClassifierWrapperContainer.propTypes = {
+  /** Callback that will be called with a subject ID when the classification subject is added to a collection. */
   onAddToCollection: func,
-  authClient: shape({}),
-  project: shape({})
+  /** optional subjectID (from the page URL.) */
+  subjectID: string,
+  /** optional subject set ID (from the page URL.) */
+  subjectSetID: string,
+  /** required workflow ID (from the page URL.) */
+  workflowID: string
 }
 
-export default withStore(ClassifierWrapperContainer)
+export default observer(ClassifierWrapperContainer)
